usb: dwc2: invalidate the dcache before starting the DMA
We should invalidate the dcache before starting the DMA. In case there are
any dirty lines from the DMA buffer in the cache, subsequent cache-line
replacements may corrupt the buffer in memory while the DMA is still going on.
Cache-line replacement can happen if the CPU tries to bring some other memory
locations into the cache while the DMA is going on.

Change-Id: Icabd174f1832eb3c5375d6b9efc6858571e588a6
Signed-off-by: Eddie Cai <eddie.cai.linux@gmail.com>
Signed-off-by: yi-hsin_hung <yi-hsin_hung@asus.com>