rockchip: rk3288: Add support for drive-strength in PINCTRL
Currently, drive-strenght to 12ma are described and supposed to be used
on RK3288. However, the pinctrl driver for this SoC only handles muxing
and pull up/pull down via PU/PD control registers. So complex IPs like
GMAC are working in normal ethernet 100mbps, but not at 1gbps typically.

This commit adds support for handling drive-strength of 12ma, when it's
defined in the DT.

Change-Id: I3a7629e9972cf41480074e14b4382b350356036c
Signed-off-by: Romain Perier <romain.perier@collabora.com>
Acked-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Reviewed-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>
Reviewed-on: https://tp-biosrd-v02/gerrit/81916
Reviewed-by: Alex Cheng(鄭富元) <Alex_Cheng@asus.com>
Tested-by: Alex Cheng(鄭富元) <Alex_Cheng@asus.com>